+++ src/helpers/dataset/getTotalExperiments.js
@@ -1,0 +1,5 @@
+import isEmptyObject from 'helpers/isEmptyObject'
+
+// Returns the total length of experiments added in My Dataset
+export default (dataset) =>
+  isEmptyObject(dataset) ? 0 : Object.keys(dataset).length
