+++ src/helpers/dataset/getTotalSamples.js
@@ -1,0 +1,6 @@
+import isEmptyObject from 'helpers/isEmptyObject'
+import unionizeArrays from 'helpers/unionizeArrays'
+
+// Returns the total length of samples added in My Dataset
+export default (dataset) =>
+  isEmptyObject(dataset) ? 0 : unionizeArrays(...Object.values(dataset)).length
